+++ src/withComponents/Footer/index.tsx
@@ -1,7 +1,8 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import clsx from 'clsx';
 import DropDownMenu from '../DropDownMenu';
-import './index.css';
+import BackgroundPicker from '../BackgroundPicker';
+import AddAppItem from '../AddAppItem';
 
 import addImg from './images/emoji_ADD.svg';
 import cancelImg from './images/emoji_OUT.svg';
@@ -10,12 +11,21 @@ import PostLink from '../PostLink';
 
 import useRoomState from '../../hooks/useRoomState/useRoomState';
 import { useWidgetContext } from '../../withHooks/useWidgetContext/useWidgetContext';
+import './index.css';
+
+import LinkIcon from '../../images/link.svg';
 
 type FooterProps = {
   classNames?: string;
 };
 
+enum Menu {
+  MENU_ROOT,
+  MENU_ADD_LINK,
+}
+
 const Footer = ({ classNames }: FooterProps) => {
+  const [menuState, setMenuState] = useState(Menu.MENU_ROOT);
   const roomState = useRoomState();
   const { addWidget } = useWidgetContext();
 
@@ -26,6 +36,41 @@ const Footer = ({ classNames }: FooterProps) => {
     setIsAddAppOpen(false);
   };
 
+  const onAddLinkClick = () => {
+    setMenuState(Menu.MENU_ADD_LINK);
+  };
+
+  const onBackBtnClick = () => {
+    setMenuState(Menu.MENU_ROOT);
+  };
+
+  const mainMenu =
+    menuState === Menu.MENU_ROOT ? (
+      <>
+        <BackgroundPicker />
+        <div className="Footer-addAppList">
+          <div className="Footer-menuTitle">Add an app</div>
+          <AddAppItem
+            imgSrc={LinkIcon}
+            imgAltText="link img icon"
+            title="Add a link"
+            descText="Link to a URL from inside the room for everybody to access."
+            onClickHandler={onAddLinkClick}
+          />
+        </div>
+      </>
+    ) : null;
+
+  const addLink =
+    menuState === Menu.MENU_ADD_LINK ? (
+      <>
+        <div className="Footer-menuBackBtn" onClick={onBackBtnClick}>
+          &lt;
+        </div>
+        <PostLink onSubmitHandler={onUrlSubmitHandler} />
+      </>
+    ) : null;
+
   return (
     <footer className={clsx('Footer', classNames)}>
       {roomState === 'disconnected' ? null : (
@@ -34,12 +79,14 @@ const Footer = ({ classNames }: FooterProps) => {
           buttonAltText={isAddAppOpen ? 'addImg' : 'cancelImg'}
           onButtonClickHandler={() => {
             setIsAddAppOpen(!isAddAppOpen);
+            setMenuState(Menu.MENU_ROOT);
           }}
           classNames="Footer-btn"
           isOpenDirectionDown={false}
           isActive={isAddAppOpen}
         >
-          <PostLink onSubmitHandler={onUrlSubmitHandler} />
+          {mainMenu}
+          {addLink}
         </DropDownMenu>
       )}
     </footer>